backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-vision-imageanalysis==1.0.0
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
blinker==1.9.0
blis==1.3.0
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.4
cymem==2.0.11
dateparser==1.2.1
Deprecated==1.2.18
distro==1.9.0
docx2txt==0.9
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.5.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
limits==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.10.1
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
msgspec==0.19.0
msrest==0.7.1
murmurhash==1.0.13
mysql-connector-python==9.3.0
numpy==2.3.1
oauthlib==3.3.1
openai==1.90.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
preshed==3.0.10
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyodbc==5.2.0
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic-bin==0.4.14
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
sse-starlette==2.3.6
starlette==0.47.1
thinc==8.3.6
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.2
